Move stylesheets (#26)
